REF: organize DataFrame tests for indexing, by-method (#31558)
