+++ pandas/tests/frame/indexing/test_datetime.py
@@ -45,13 +45,6 @@ class TestDataFrameIndexingDatetimeWithTZ:
         df = result.set_index("foo")
         tm.assert_index_equal(df.index, idx)
 
-    def test_transpose(self, timezone_frame):
-
-        result = timezone_frame.T
-        expected = DataFrame(timezone_frame.values.T)
-        expected.index = ["A", "B", "C"]
-        tm.assert_frame_equal(result, expected)
-
     def test_scalar_assignment(self):
         # issue #19843
         df = pd.DataFrame(index=(0, 1, 2))
