+++ pandas/tests/frame/indexing/test_iat.py
@@ -1,0 +1,7 @@
+def test_iat(float_frame):
+
+    for i, row in enumerate(float_frame.index):
+        for j, col in enumerate(float_frame.columns):
+            result = float_frame.iat[i, j]
+            expected = float_frame.at[row, col]
+            assert result == expected
