+++ pandas/tests/frame/indexing/test_where.py
@@ -10,22 +10,30 @@ from pandas import DataFrame, DatetimeIndex, Series, Timestamp, date_range, isna
 import pandas._testing as tm
 
 
+@pytest.fixture(params=["default", "float_string", "mixed_float", "mixed_int"])
+def where_frame(request, float_string_frame, mixed_float_frame, mixed_int_frame):
+    if request.param == "default":
+        return DataFrame(np.random.randn(5, 3), columns=["A", "B", "C"])
+    if request.param == "float_string":
+        return float_string_frame
+    if request.param == "mixed_float":
+        return mixed_float_frame
+    if request.param == "mixed_int":
+        return mixed_int_frame
+
+
+def _safe_add(df):
+    # only add to the numeric items
+    def is_ok(s):
+        return (
+            issubclass(s.dtype.type, (np.integer, np.floating)) and s.dtype != "uint8"
+        )
+
+    return DataFrame(dict((c, s + 1) if is_ok(s) else (c, s) for c, s in df.items()))
+
+
 class TestDataFrameIndexingWhere:
-    def test_where(self, float_string_frame, mixed_float_frame, mixed_int_frame):
-        default_frame = DataFrame(np.random.randn(5, 3), columns=["A", "B", "C"])
-
-        def _safe_add(df):
-            # only add to the numeric items
-            def is_ok(s):
-                return (
-                    issubclass(s.dtype.type, (np.integer, np.floating))
-                    and s.dtype != "uint8"
-                )
-
-            return DataFrame(
-                dict((c, s + 1) if is_ok(s) else (c, s) for c, s in df.items())
-            )
-
+    def test_where_get(self, where_frame, float_string_frame):
         def _check_get(df, cond, check_dtypes=True):
             other1 = _safe_add(df)
             rs = df.where(cond, other1)
@@ -40,19 +48,15 @@ class TestDataFrameIndexingWhere:
                 assert (rs.dtypes == df.dtypes).all()
 
         # check getting
-        for df in [
-            default_frame,
-            float_string_frame,
-            mixed_float_frame,
-            mixed_int_frame,
-        ]:
-            if df is float_string_frame:
-                with pytest.raises(TypeError):
-                    df > 0
-                continue
-            cond = df > 0
-            _check_get(df, cond)
+        df = where_frame
+        if df is float_string_frame:
+            with pytest.raises(TypeError):
+                df > 0
+            return
+        cond = df > 0
+        _check_get(df, cond)
 
+    def test_where_upcasting(self):
         # upcasting case (GH # 2794)
         df = DataFrame(
             {
@@ -78,6 +82,7 @@ class TestDataFrameIndexingWhere:
 
         tm.assert_series_equal(result, expected)
 
+    def test_where_alignment(self, where_frame, float_string_frame):
         # aligning
         def _check_align(df, cond, other, check_dtypes=True):
             rs = df.where(cond, other)
@@ -107,27 +112,30 @@ class TestDataFrameIndexingWhere:
             if check_dtypes and not isinstance(other, np.ndarray):
                 assert (rs.dtypes == df.dtypes).all()
 
-        for df in [float_string_frame, mixed_float_frame, mixed_int_frame]:
-            if df is float_string_frame:
-                with pytest.raises(TypeError):
-                    df > 0
-                continue
+        df = where_frame
+        if df is float_string_frame:
+            with pytest.raises(TypeError):
+                df > 0
+            return
 
-            # other is a frame
-            cond = (df > 0)[1:]
-            _check_align(df, cond, _safe_add(df))
+        # other is a frame
+        cond = (df > 0)[1:]
+        _check_align(df, cond, _safe_add(df))
 
-            # check other is ndarray
-            cond = df > 0
-            _check_align(df, cond, (_safe_add(df).values))
+        # check other is ndarray
+        cond = df > 0
+        _check_align(df, cond, (_safe_add(df).values))
 
-            # integers are upcast, so don't check the dtypes
-            cond = df > 0
-            check_dtypes = all(not issubclass(s.type, np.integer) for s in df.dtypes)
-            _check_align(df, cond, np.nan, check_dtypes=check_dtypes)
+        # integers are upcast, so don't check the dtypes
+        cond = df > 0
+        check_dtypes = all(not issubclass(s.type, np.integer) for s in df.dtypes)
+        _check_align(df, cond, np.nan, check_dtypes=check_dtypes)
 
+    def test_where_invalid(self):
         # invalid conditions
-        df = default_frame
+        df = DataFrame(np.random.randn(5, 3), columns=["A", "B", "C"])
+        cond = df > 0
+
         err1 = (df + 1).values[0:2, :]
         msg = "other must be the same shape as self when an ndarray"
         with pytest.raises(ValueError, match=msg):
@@ -144,7 +152,9 @@ class TestDataFrameIndexingWhere:
         with pytest.raises(ValueError, match=msg):
             df.mask(0)
 
+    def test_where_set(self, where_frame, float_string_frame):
         # where inplace
+
         def _check_set(df, cond, check_dtypes=True):
             dfi = df.copy()
             econd = cond.reindex_like(df).fillna(True)
@@ -160,27 +170,23 @@ class TestDataFrameIndexingWhere:
                         v = np.dtype("float64")
                     assert dfi[k].dtype == v
 
-        for df in [
-            default_frame,
-            float_string_frame,
-            mixed_float_frame,
-            mixed_int_frame,
-        ]:
-            if df is float_string_frame:
-                with pytest.raises(TypeError):
-                    df > 0
-                continue
+        df = where_frame
+        if df is float_string_frame:
+            with pytest.raises(TypeError):
+                df > 0
+            return
 
-            cond = df > 0
-            _check_set(df, cond)
+        cond = df > 0
+        _check_set(df, cond)
 
-            cond = df >= 0
-            _check_set(df, cond)
+        cond = df >= 0
+        _check_set(df, cond)
 
-            # aligning
-            cond = (df >= 0)[1:]
-            _check_set(df, cond)
+        # aligning
+        cond = (df >= 0)[1:]
+        _check_set(df, cond)
 
+    def test_where_series_slicing(self):
         # GH 10218
         # test DataFrame.where with Series slicing
         df = DataFrame({"a": range(3), "b": range(4, 7)})
